src/mcp_salesforce/pyproject.toml
[project]
name = "mcp_salesforce"
version = "0.1.0"
description = "MCP Salesforce service"
authors = [{ name = "Jaroslav" }]
dependencies = [
    "fastapi",
    "fastmcp",
    "python-dotenv",
    "simple-salesforce",
    "mcp_common @ file:///src/mcp_common"
]
requires-python = ">=3.10"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
